+++ src/main/java/com/demo/entrepreneur/dto/user/RequestUserDto.java
@@ -1,29 +1,35 @@
-package com.demo.entrepreneur.dto;
+package com.demo.entrepreneur.dto.user;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.AccessMode;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class ResponseUserDto {
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class RequestUserDto {
 
-    @Schema(description = "user login", required = true, accessMode = AccessMode.READ_ONLY)
+    @Schema(description = "user login", required = true)
     @JsonProperty("login")
     private String login;
+
+    @Schema(description = "user password", required = true, accessMode = AccessMode.WRITE_ONLY)
+    @JsonProperty("password")
+    private String password;
+
     @Schema(required = true)
     @JsonProperty("email")
     private String email;
+
     @Schema(required = true)
     @JsonProperty("firstName")
     private String firstName;
+
     @Schema(required = true)
     @JsonProperty("lastName")
     private String lastName;
