+++ src/main/java/com/demo/entrepreneur/dto/user/ResponseUserDto.java
@@ -1,35 +1,29 @@
-package com.demo.entrepreneur.dto;
+package com.demo.entrepreneur.dto.user;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.AccessMode;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class RequestUserDto {
+public class ResponseUserDto {
 
-    @Schema(description = "user login", required = true)
+    @Schema(description = "user login", required = true, accessMode = AccessMode.READ_ONLY)
     @JsonProperty("login")
     private String login;
-
-    @Schema(description = "user password", required = true, accessMode = AccessMode.WRITE_ONLY)
-    @JsonProperty("password")
-    private String password;
-
     @Schema(required = true)
     @JsonProperty("email")
     private String email;
-
     @Schema(required = true)
     @JsonProperty("firstName")
     private String firstName;
-
     @Schema(required = true)
     @JsonProperty("lastName")
     private String lastName;
