+++ src/main/java/com/demo/entrepreneur/service/impl/DefaultUserService.java
@@ -2,7 +2,7 @@ package com.demo.entrepreneur.service.impl;
 
 import com.demo.entrepreneur.exception.UnsupportedEmailException;
 import com.demo.entrepreneur.repository.UserRepository;
-import com.demo.entrepreneur.dto.RequestUserDto;
+import com.demo.entrepreneur.dto.user.RequestUserDto;
 import com.demo.entrepreneur.entity.User;
 import com.demo.entrepreneur.mapping.populator.impl.RequestUserPopulator;
 import com.demo.entrepreneur.service.EmailSenderService;
