+++ src/main/resources/application.properties
@@ -12,7 +12,7 @@ cron.template.exchangeRate= 0 0 11 * * ?
 archive.load.rate.millis=10000
 
 #yyyy-MM-dd
-archive.date.begin=2019-10-01
+archive.date.begin=2015-01-01
 archive.date.api.format=dd.MM.YYYY
 #API URLs
 exchangeRate.api.url=https://api.privatbank.ua/p24api/pubinfo?json&exchange&coursid=3
